requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
pyproj>=3.6.0
python-docx>=1.1.0
pdfplumber>=0.11.0